change details
side flow
